+++ tests/Doka.EntityFrameworkCore.SafeMigrations.PostgreSql.Tests/Imports.cs
@@ -10,6 +10,7 @@ global using System.Text;
 global using System.Threading;
 global using System.Threading.Tasks;
 global using Doka.EntityFrameworkCore.SafeMigrations.Testing;
+global using FsCheck.Xunit;
 global using Microsoft.EntityFrameworkCore;
 global using Microsoft.EntityFrameworkCore.Design;
 global using Microsoft.EntityFrameworkCore.Infrastructure;